Extracting out the server context to a top level class.
